requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.7
numpy==1.26.4
pyepics==3.5.2
pyparsing==3.1.1
requests==2.32.3
urllib3==2.2.2
